inventory/app/requirements.txt
flask>=3.0
pyyaml>=6.0
openai>=1.0
pyserial>=3.5
# Optional: for agent device web search (GET /api/agent/device-search)
duckduckgo-search>=6.0
